Add functioning(?) similarity matching to catalog catalog_search
It's not clear whether this works right now. If it works,
then we need to combine the vectors of summaries from the
catalog in a different manner in order to get accurate results.

If the large number of matches is inaccurate, we need to confirm debug
to find figure out expected results.